Перенесла методы по обработке сообщений из AdsController в CommentController
