+++ src/main/java/ru/skypro/homework/controller/AdsController.java
@@ -100,58 +100,4 @@ public class AdsController {
 
         return ResponseEntity.ok(image);
     }
-
-    @GetMapping("/{id}/comments")
-    public ResponseEntity<ResponseWrapperComment> getComments(@PathVariable(name = "id") int id) {
-        if (false) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
-        return ResponseEntity.ok(new ResponseWrapperComment());
-    }
-
-    @PostMapping("/{id}/comments")
-    public ResponseEntity<Comment> sendComment(@PathVariable(name = "id") int id,
-                                               @RequestBody Comment comment) {
-        if (false) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-        if (false) {
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-        }
-        if (false) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
-        return ResponseEntity.ok(comment);
-    }
-
-    @DeleteMapping("/{adId}/comments/{commentId}")
-    public ResponseEntity<Void> deleteComment(@PathVariable(name = "adId") int adId,
-                                              @PathVariable(name = "commentId") int commentId) {
-        if (false) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-        if (false) {
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-        }
-        if (false) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
-        return ResponseEntity.ok().build();
-    }
-
-    @PatchMapping("/{adId}/comments/{commentId}")
-    public ResponseEntity<Comment> editComment(@PathVariable(name = "adId") int adId,
-                                               @PathVariable(name = "commentId") int commentId,
-                                               @RequestBody Comment comment){
-        if (false) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-        if (false) {
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-        }
-        if (false) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
-        return ResponseEntity.ok(comment);
-    }
 }
