+++ src/main/java/ru/skypro/homework/controller/CommentController.java
@@ -1,0 +1,67 @@
+package ru.skypro.homework.controller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import ru.skypro.homework.dto.Comment;
+import ru.skypro.homework.dto.ResponseWrapperComment;
+
+@CrossOrigin(value = "http://localhost:3000")
+@RestController
+@RequestMapping("/ads")
+public class CommentController {
+
+    @GetMapping("/{id}/comments")
+    public ResponseEntity<ResponseWrapperComment> getComments(@PathVariable(name = "id") int id) {
+        if (false) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return ResponseEntity.ok(new ResponseWrapperComment());
+    }
+
+    @PostMapping("/{id}/comments")
+    public ResponseEntity<Comment> sendComment(@PathVariable(name = "id") int id,
+                                               @RequestBody Comment comment) {
+        if (false) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (false) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+        if (false) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return ResponseEntity.ok(comment);
+    }
+
+    @DeleteMapping("/{adId}/comments/{commentId}")
+    public ResponseEntity<Void> deleteComment(@PathVariable(name = "adId") int adId,
+                                              @PathVariable(name = "commentId") int commentId) {
+        if (false) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (false) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+        if (false) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return ResponseEntity.ok().build();
+    }
+
+    @PatchMapping("/{adId}/comments/{commentId}")
+    public ResponseEntity<Comment> editComment(@PathVariable(name = "adId") int adId,
+                                               @PathVariable(name = "commentId") int commentId,
+                                               @RequestBody Comment comment) {
+        if (false) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (false) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+        if (false) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return ResponseEntity.ok(comment);
+    }
+}
